requirements.txt
absl-py==2.1.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astunparse==1.6.3
bert-embedding==1.0.1
black==21.4b2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==3.1.0
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
detectron2==0.6+cu113
einops==0.8.0
et_xmlfile==2.0.0
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.12.0
ftfy==6.2.3
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.4.0
gluonnlp==0.6.0
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.68.1
h5py==3.11.0
hausdorff==0.2.6
huggingface-hub==0.27.0
hydra-core==1.3.2
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.9
joblib @ file:///croot/joblib_1718217211762/work
keras==2.13.1
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.41.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
ml-collections==0.1.1
mmcv-full==1.3.12
mmengine==0.10.5
mmsegmentation==0.17.0
MultiScaleDeformableAttention==1.0
mxnet==1.4.0
mypy-extensions==1.0.0
networkx==3.1
nltk==3.9.1
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1701735466804/work
omegaconf==2.3.0
openai-clip==1.0.1
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
pathspec==0.12.1
Pillow==9.5.0
platformdirs==4.3.6
portalocker==3.0.0
prettytable==3.11.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
pydot==3.0.3
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn @ file:///croot/scikit-learn_1694788527225/work
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.6
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
textblob==0.18.0.post0
thop==0.1.1.post2209072238
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
tifffile==2023.7.10
timm==1.0.12
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.0
tqdm==4.67.1
transformers==4.46.3
typing==3.6.6
typing_extensions==4.5.0
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
wrapt==1.17.0
yacs==0.1.8
yapf==0.43.0
zipp==3.20.2
